requirements.txt
matplotlib
matplotlib-inline
numpy
opencv-python
pillow
torch
torchvision
tensorboard
segment_anything
transformers